# docs/swagger.yaml
definitions:
  model.AuthenticationInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  model.AuthenticationResponse:
    type: object
  model.CreateSubscriptionInput:
    properties:
      description:
        type: string
      subscriptionName:
        type: string
    type: object
  model.Email:
    properties:
      body:
        type: string
      subject:
        type: string
    type: object
  model.MessageResponse:
    properties:
      message:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: milaguyenlam@gmail.com
    name: Nguyen Thanh Lam
    url: https://github.com/milaguyenlam
  description: This is a newsletter subscription API service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: STRV Newsletter Subscription API
  version: "1.0"
paths:
  /subscription/{subscriptionID}/send:
    post:
      consumes:
      - application/json
      description: Send a newsletter email to all subscribers
      operationId: send-newsletter
      parameters:
      - description: Subscription ID
        in: path
        name: subscriptionID
        required: true
        type: string
      - description: Email details
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/model.Email'
      produces:
      - application/json
      responses:
        "200":
          description: Message
          schema:
            $ref: '#/definitions/model.MessageResponse'
        "500":
          description: Error message
          schema:
            $ref: '#/definitions/model.MessageResponse'
      security:
      - Bearer: []
      summary: Send a newsletter
  /subscription/{subscriptionID}/subscribe:
    get:
      description: Subscribe to a newsletter with a given subscription ID (confirmation
        email with unsubscribe link will be sent to newly subscribed user)
      operationId: subscribe-newsletter
      parameters:
      - description: Subscription ID
        in: path
        name: subscriptionID
        required: true
        type: string
      - description: Email to subscribe
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Message
          schema:
            $ref: '#/definitions/model.MessageResponse'
        "500":
          description: Error message
          schema:
            $ref: '#/definitions/model.MessageResponse'
      summary: Subscribe to a newsletter
  /subscription/{subscriptionID}/unsubscribe:
    get:
      description: Unsubscribe from a newsletter with a given subscription ID
      operationId: unsubscribe-newsletter
      parameters:
      - description: Subscription ID
        in: path
        name: subscriptionID
        required: true
        type: string
      - description: Email to unsubscribe
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Message
          schema:
            $ref: '#/definitions/model.MessageResponse'
        "500":
          description: Error message
          schema:
            $ref: '#/definitions/model.MessageResponse'
      summary: Unsubscribe from a newsletter
  /subscription/create:
    post:
      consumes:
      - application/json
      description: Create a new subscription with the given name and description
      operationId: create-subscription
      parameters:
      - description: Subscription input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CreateSubscriptionInput'
      produces:
      - application/json
      responses:
        "200":
          description: Subscription ID - composed of its name and the editor's email
            divided by underscore.
          schema:
            $ref: '#/definitions/model.MessageResponse'
        "500":
          description: Error message
          schema:
            $ref: '#/definitions/model.MessageResponse'
      security:
      - Bearer: []
      summary: Create a new subscription
  /user/login:
    post:
      consumes:
      - application/json
      description: Logs in a user
      operationId: login
      parameters:
      - description: Login credentials
        in: body
        name: authInput
        required: true
        schema:
          $ref: '#/definitions/model.AuthenticationInput'
      produces:
      - application/json
      responses:
        "200":
          description: Token
          schema:
            $ref: '#/definitions/model.AuthenticationResponse'
        "401":
          description: Message
          schema:
            $ref: '#/definitions/model.MessageResponse'
      summary: User Login
  /user/register:
    post:
      consumes:
      - application/json
      description: Registers a new user (user email has to be verified by AWS SES
        to be able to send newsletter emails)
      operationId: register
      parameters:
      - description: Registration details
        in: body
        name: authInput
        required: true
        schema:
          $ref: '#/definitions/model.AuthenticationInput'
      produces:
      - application/json
      responses:
        "200":
          description: Token
          schema:
            $ref: '#/definitions/model.AuthenticationResponse'
        "401":
          description: Message
          schema:
            $ref: '#/definitions/model.MessageResponse'
      summary: User Registration
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"